Begin fleshing out what the API client will look like.
We have a couple goals with the API client at the moment.

1. Don't be required unless explicitly wanted
The gem is currently really simple and the functionality is very
minimal. Most people won't need to interact with the API so it should be
opt in.

2. Don't depend on any outside dependencies
Pretty much every application depends on some HTTP client like Faraday
or Excon. Net::HTTP is available everywhere and isn't _that_ painful to
use.