Improve get_key error handling. Add Snitcher::API::Error.
The error cases for `get_key` need improvement. `nil` was returned if
any error occurred on the server. A timeout returned a hash with a
single message key.

I have introduced a standard error class that will be used for any error
from the API that should raise an exception. I have decided to deviate
from the naming used for the "type" returned as I don't find all of them
that helpful in the context of an application.