Refactor Snitcher::API::Client.
I have tried to remove some indirection and duplication in the client. I
have removed the Base class as we do not currently plan to reuse that
logic. It did serve as a nice way to separate the logic of making
requests from defining what requests were allowed. I have been able to
remove enough duplication that the code to make requests is currently
fairly small.

Timeout::Error is now raised when a timeout occurs. This is more
idomatic Ruby.

All requests with a body are now considered to be JSON as the API itself
only supports JSON. Should we add other methods in the future we'll
review it again then.

There is more refactoring coming but I want to isolate those into
separate commits.