Change `Snitcher::API::Client` to take the key as a param.
The API key was a required parameter of the client and hiding it inside
of an options hash makes it possible to create a client without knowing
you have to set the api key. I believe that a required parameter should
be just that, a parameter and not an option in a hash.