Support spaces in tags when filtering snitches by tags.
We were doing a nieve join of the given array into the query parameter
but instead we need to make sure that the query is properly url encoded.

I've changed the API to use a vararg for filtering snitches as it will
make using the API easier. It also supports passing in an array
(possibly multiple arrays) of tags with which to filter by.

Fixes #15.