Raise specific errors for most failure cases.
When making API call there are a number of errors that can occur. From
things like credentials being invalid to problems with finding snitches
there are a number of things that can go wrong. This changes the error
handling from returning a Hash in some cases to raising granular
exceptions.