Convert documentation to Yardoc. Documentation pass.
While everything was reasonably documented with Tomdoc the big advantage
to Yardoc is that it is processable by a computer. This means that
people using IDEs can get useful information about types and expected
return values.

I've also taken this chance to better document how errors are expected
to be raised and review what return values are.