Trim Client's API. Remove clear_tags and replace_tags.
Both of these are effectively helper methods for calls to `edit_snitch`.
While I'm a big fan of usability in an API design I would rather add
these scenarios later on when there is a clear need for them. My concern
right now as we get closer to a release is to have a more concentrated
and focused API.

That said I am not sure about `add_tags` or `remove_tag` as Client API
methods. The main reason I am leaving them in is that they are exposed
in the Dead Man's Snitch API and not just helpers on top of
functionality provided by other means.

Finally, I've added examples to `edit_snitch` for clearing and removing
tags. Additionally I've tested a couple cases that do add some usability
improvements for users.