Add :interval as top level attribute for create and edit.
While the Dead Man's Snitch API requires specifying the interval inside
of a type object (e.g. `{ "type": { "interval": "hourly" }}`) it can be
confusing for users. This is compounded by the fact that there are no
other properties other than `interval` inside of type.

The original design intent was to allow for multiple types of Snitches,
which is something that will likely happen. At least with current
thinking `interval` will likely always be required.

Right now it's a usability win and we can always do a version bump and
deprecate it.