Remove Snitcher::API.get_key.
Wait... what!? While retrieving the API key given a username and
password is part of the API, I expect we will be deprecating it in the
near future. The reason is that it's not very flexible. There are some
things on the roadmap (individual sessions, teams, etc...) that will
need a more flexible approach.

I think that it is better to remove it now rather than break customer's
applications when we deprecate or remove it. Worst case, we can always
reintroduce it in the future.