Use bundler 1.x for older versions of Ruby and remove 1.9
Bundler 2.x requires Ruby 2.3+ but we want to be able to support older
versions back to at least 2.0. I've had to remove 1.9 from the test
matrix since it is failing due to the JSON gem no longer being
compatible. This locks to the latest version of Bundler 1.x for those
versions of Ruby that don't support 2.0.

Ruby 2.0 and 2.1 are not available on the Bionic images.